theme updater

+++ run_update.sh
@@ -68,33 +68,88 @@ echo "/bin/bash ~/startup.sh" > ~/.bashrc
 echo -e "${GREEN}OK${NONE}"
 
 echo -ne  "Updating Bundled Themes (won't affect any other themes)..."
-cd ~/storage/shared/pegasus-frontend/themes/RP-epic-noir
-git reset --hard &> ~/storage/shared/pegasus_installer_log.log 
-git pull &> ~/storage/shared/pegasus_installer_log.log 
 
-cd ~/storage/shared/pegasus-frontend/themes/RP-switch
-git reset --hard &> ~/storage/shared/pegasus_installer_log.log 
-git pull &> ~/storage/shared/pegasus_installer_log.log 
+FOLDER=~/storage/shared/pegasus-frontend/themes/RP-epic-noir
+if [ -d "$FOLDER" ]; then
+	cd ~/storage/shared/pegasus-frontend/themes/RP-epic-noir
+	git reset --hard &> ~/storage/shared/pegasus_installer_log.log 
+	git pull &> ~/storage/shared/pegasus_installer_log.log 
+	
+else
+	echo -ne "Downloading Pegasus Theme : RP Epic Noir..."
+	git clone https://github.com/dragoonDorise/RP-epic-noir.git ~/storage/shared/pegasus-frontend/themes/RP-epic-noir &>> ~/storage/shared/pegasus_installer_log.log
+	echo -e "${GREEN}OK${NONE}"
+fi
+
+FOLDER=~/storage/shared/pegasus-frontend/themes/RP-switch
+if [ -d "$FOLDER" ]; then
+	cd ~/storage/shared/pegasus-frontend/themes/RP-switch
+	git reset --hard &> ~/storage/shared/pegasus_installer_log.log 
+	git pull &> ~/storage/shared/pegasus_installer_log.log 
+else
+	echo -ne "Downloading Pegasus Theme : RP Switch..."
+	#We delete the theme, for previous users
+	rm -rf ~/storage/shared/pegasus-frontend/themes/RP-switch &>> ~/storage/shared/pegasus_installer_log.log
+	git clone https://github.com/dragoonDorise/RP-switch.git ~/storage/shared/pegasus-frontend/themes/RP-switch &>> ~/storage/shared/pegasus_installer_log.log
+	echo -e "${GREEN}OK${NONE}"
+fi
+
 
 if [ $handheldModel != 'RP2+' ]; then
-	cd ~/storage/shared/pegasus-frontend/themes/gameOS
-	git reset --hard &> ~/storage/shared/pegasus_installer_log.log 
-	git pull &> ~/storage/shared/pegasus_installer_log.log 
 	
+	FOLDER=~/storage/shared/pegasus-frontend/themes/gameOS
+	if [ -d "$FOLDER" ]; then
+		cd ~/storage/shared/pegasus-frontend/themes/gameOS
+		git reset --hard &> ~/storage/shared/pegasus_installer_log.log 
+		git pull &> ~/storage/shared/pegasus_installer_log.log 
+		
+	else
+		echo -ne "Downloading Pegasus Theme : GameOS..."
+		git clone https://github.com/PlayingKarrde/gameOS.git ~/storage/shared/pegasus-frontend/themes/gameOS &>> ~/storage/shared/pegasus_installer_log.log
+		echo -e "${GREEN}OK${NONE}"
+	fi
 	
-	cd ~/storage/shared/pegasus-frontend/themes/clearOS
-	git reset --hard &> ~/storage/shared/pegasus_installer_log.log 
-	git pull &> ~/storage/shared/pegasus_installer_log.log 
+	FOLDER=~/storage/shared/pegasus-frontend/themes/clearOS
+	if [ -d "$FOLDER" ]; then
+		cd ~/storage/shared/pegasus-frontend/themes/clearOS
+		git reset --hard &> ~/storage/shared/pegasus_installer_log.log 
+		git pull &> ~/storage/shared/pegasus_installer_log.log 
+		
+	else
+		echo -ne "Downloading Pegasus Theme : ClearOS..."
+		git clone https://github.com/PlayingKarrde/clearOS.git ~/storage/shared/pegasus-frontend/themes/clearOS &>> ~/storage/shared/pegasus_installer_log.log
+		echo -e "${GREEN}OK${NONE}"
+	fi
 	
-	
-	cd ~/storage/shared/pegasus-frontend/themes/neoretro-dark
-	git reset --hard &> ~/storage/shared/pegasus_installer_log.log 
-	git pull &> ~/storage/shared/pegasus_installer_log.log 
+	FOLDER=~/storage/shared/pegasus-frontend/themes/neoretro-dark
+	if [ -d "$FOLDER" ]; then
+		cd ~/storage/shared/pegasus-frontend/themes/neoretro-dark
+		git reset --hard &> ~/storage/shared/pegasus_installer_log.log 
+		git pull &> ~/storage/shared/pegasus_installer_log.log 
+		
+	else
+		echo -ne "Downloading Pegasus Theme : NeoRetro Dark..."
+		git clone https://github.com/TigraTT-Driver/neoretro-dark.git ~/storage/shared/pegasus-frontend/themes/neoretro-dark &>> ~/storage/shared/pegasus_installer_log.log
+		echo -e "${GREEN}OK${NONE}"
+	fi
+
 fi
+
+#RP2 Exclusive themes
 if [ $handheldModel == 'RP2+' ]; then
+
+	FOLDER=~/storage/shared/pegasus-frontend/themes/neoretro-dark
+	if [ -d "$FOLDER" ]; then
 	cd ~/storage/shared/pegasus-frontend/themes/retromega
 	git reset --hard &> ~/storage/shared/pegasus_installer_log.log 
 	git pull &> ~/storage/shared/pegasus_installer_log.log 
+		
+	else
+		echo -ne "Downloading Pegasus Theme : Retro Mega..."
+		git clone https://github.com/djfumberger/retromega.git ~/storage/shared/pegasus-frontend/themes/retromega &>> ~/storage/shared/pegasus_installer_log.log
+		echo -e "${GREEN}OK${NONE}"
+	fi
+
 fi
 echo -e "${GREEN}OK${NONE}"
